pyproject.toml
[project]
name = "arxiv-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.16",
    "dotenv>=0.9.9",
    "kaggle>=1.7.4.5",
    "opensearch-py>=3.0.0",
    "pandas==2.1.4",
]

[tool.ruff]
exclude = ["*.venv*"]
line-length = 120

[tool.ruff.lint.isort] # https://docs.astral.sh/ruff/settings/#lintisort
combine-as-imports = false             # 預設不合併 `as` 別名的 import
force-single-line = false              # 預設不強制每個 import 獨立成一行 (不管長度)
known-first-party = ["app", "de_*kit"]

[dependency-groups]
deploy = ["aws-cdk-lib>=2.212.0", "constructs>=10.0.0,<11.0.0"]
dev = ["ipykernel>=6.30.1", "pytest==6.2.5"]


[tool.ruff.lint] # https://docs.astral.sh/ruff/settings/#lint
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
